requirements.txt
azure-cognitiveservices-speech>=1.36.0
google-cloud-speech>=2.24.0
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.31.0
openai>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
google-cloud-storage>=2.0.0
